requirements.txt

fastapi==0.110.0
uvicorn==0.29.0

qiskit==1.0.2
qiskit-aer==0.14.2
qiskit-ibm-runtime==0.22.0

pyquil==4.12.0
cirq==1.3.0
cirq-ionq==1.3.0
cirq-google==1.3.0

amazon-braket-sdk==1.83.0

torch==2.2.1
transformers==4.39.3
bitsandbytes==0.43.0
accelerate==0.28.0

apscheduler==3.10.4
networkx==3.2.1
pydantic==2.6.4
httpx==0.27.0
python-dotenv==1.0.1

pytest==8.1.1
pytest-asyncio==0.23.6